chore: add maxlength to fields

+++ database/migrations/2021_10_28_053401_create_transaction_items_table.php
@@ -18,7 +18,7 @@ class CreateTransactionItemsTable extends Migration
             $table->bigInteger('users_id');
             $table->bigInteger('products_id');
             $table->bigInteger('transactions_id');
-            $table->bigInteger('quantity');
+            $table->integer('quantity');
 
             $table->timestamps();
         });
